Parameterize client secret
If a non-Google authentication provider is used, the client secret
stored in the OKpy database will be different from this hard-coded
value which causes the ok-client to ok-server communication to fail.

A work-around for this is to manually adjust the client secret in the
ok-server database via the following command.

```sql
UPDATE client
SET client_secret='EWKtcCp5nICeYgVyCPypjs3aLORqQ3H'
WHERE client_id='ok-client';
```

However, a more long-term manageable fix is to parameterize the client
secret and set it to the correct value for the authentication provider
in each ok-client deployment.